Split LEDs into submodule
