+++ src/bin/app.rs
@@ -7,7 +7,7 @@ use pico_probe as _;
 #[rtic::app(device = rp2040_hal::pac, dispatchers = [XIP_IRQ, CLOCKS_IRQ])]
 mod app {
     use core::mem::MaybeUninit;
-    use pico_probe::setup::*;
+    use pico_probe::{leds::BoardLeds, setup::*};
     use rp2040_hal::usb::UsbBus;
     use usb_device::class_prelude::*;
