+++ src/lib.rs
@@ -9,6 +9,7 @@ use rtic_monotonics::{rp2040::Timer, Monotonic};
 
 pub mod dap;
 pub mod device_signature;
+pub mod leds;
 pub mod pio;
 pub mod setup;
 pub mod systick_delay;
